+++ foxes/input/states/__init__.py
@@ -7,5 +7,6 @@ from .scan_ws import ScanWS
 from .states_table import StatesTable, Timeseries, TabStates
 from .field_data_nc import FieldDataNC
 from .multi_height import MultiHeightStates, MultiHeightTimeseries
+from .multi_height import MultiHeightNCStates, MultiHeightNCTimeseries
 
 from . import create